style: highlighted labels on bar charts
Previously it was hard to tell which day was currently selected if you
were zoomed in on the bar chart and there were no hours for the
currently selected day. The labels are now styled to match the bar
colors to they denote the currently selected date even in the absence of
the bar chart.